+++ tools/src/tester/helpers.ts
@@ -9,7 +9,7 @@
 
 import { type Evaluation, Result } from './types/eval.types'
 
-export function overall_result (evaluations: Evaluation[]): Result {
+export function overall_result(evaluations: Evaluation[]): Result {
   if (evaluations.some(e => e.result === Result.ERROR)) return Result.ERROR
   if (evaluations.some(e => e.result === Result.FAILED)) return Result.FAILED
   if (evaluations.some(e => e.result === Result.SKIPPED)) return Result.SKIPPED